refactor(7092): isolate settlement flow and transport tests
The funded rehearsal now separates preparation from release reconciliation, and settlement transport evidence construction and tests live in focused modules. Touched production files are back below the repository's 200-line limit.

Constraint: Preserve the proven devnet transaction and reconciliation behavior.

Confidence: high

Scope-risk: narrow

Directive: Keep orchestration helpers single-purpose and tests outside production-size budgets.

Tested: Targeted node and harness contracts, strict clippy, formatting, and diff checks.

Not-tested: No additional funded transfer was needed for this behavior-preserving split.
