+++ crates/kamn-node/src/service_api_endpoint/live_settlement_dispatch/transport_tests.rs
@@ -1,0 +1,8 @@
+use super::*;
+
+#[test]
+fn false_confirmation_is_an_ambiguous_outcome() {
+    let error = require_confirmation(false, "finalized").expect_err("ambiguous result");
+
+    assert!(error.starts_with("SETTLEMENT_OUTCOME_AMBIGUOUS"));
+}
